requirements.txt
astroid==2.2.5
gphoto2==1.9.0
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
pylint==2.3.1
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
